link all dashboard components using Navlink from router, add activeClassName to
